+++ SuggestionProviderInterface.php
@@ -1,0 +1,29 @@
+<?php
+
+
+namespace Symfony\Cmf\Bundle\SeoBundle;
+
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Route;
+use Symfony\Component\Routing\RouteCollection;
+
+/**
+ * Interface for all suggestion providers.
+ *
+ * Those providers are responsible to create a list of suggestions (routes)
+ * for pages to visit instead of a blank 404 page.
+ *
+ * @author Maximilian Berghoff <Maximilian.Berghoff@gmx.de>
+ */
+interface SuggestionProviderInterface
+{
+    /**
+     * Based on the current request this method creates a list
+     * of suggestions as an array of routes.
+     *
+     * @param Request $request
+     *
+     * @return array|Route[]
+     */
+    public function create(Request $request);
+}